requirements.txt
# CSSM (Cepstral State Space Model)
# Requirements for NVIDIA B200 / H200 / H100 GPUs (CUDA 12.x)
# Tested with Python 3.10+

# Core JAX/Flax stack with CUDA 12 support
# B200/Blackwell requires very recent JAX (0.4.35+)
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax[cuda12]>=0.4.35
jaxlib>=0.4.35
flax>=0.8.0
optax>=0.2.0

# Checkpointing
orbax-checkpoint>=0.5.0

# Data loading
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scipy>=1.10.0

# Logging and visualization
wandb>=0.16.0
tqdm>=4.65.0
matplotlib>=3.7.0

# TensorFlow for optimized tf.data loaders (ImageNet)
tensorflow>=2.15.0
